[MCUX-87602] Fix MISRA bitwise operation issue in work queue scheduling
Address MISRA-C compliance issue in k_work_schedule_for_queue() function
where bitwise NOT operation was applied to an enum constant.

Key improvements:
- Add explicit cast to uint32_t for K_WORK_RUNNING before bitwise NOT operation
- Ensures bitwise operations are performed on unsigned integer types as required by MISRA
- Prevents implicit type conversion warnings in static analysis

This change maintains functional correctness while ensuring compliance with
MISRA-C rules for bitwise operations on enumeration types.

+++ source/porting/work/work.c
@@ -623,7 +623,7 @@ int k_work_schedule_for_queue(struct k_work_q *queue,
 	k_spinlock_key_t key = k_spin_lock(&lock);
 
 	/* Schedule the work item if it's idle or running. */
-	if (((flags_get(&work->flags) & K_WORK_MASK) & (uint32_t)(~K_WORK_RUNNING)) == 0U) {
+	if (((flags_get(&work->flags) & K_WORK_MASK) & (uint32_t)(~((uint32_t)K_WORK_RUNNING))) == 0U) {
 		if (K_TIMEOUT_EQ(delay, K_NO_WAIT))
 		{
 			ret = k_work_submit_to_queue(queue, work);
